requirements.txt
numpy==1.21.6
Pillow==7.1.2
streamlit==1.11.1
tensorflow==2.8.2